+++ packages/powerscore/src/index.ts
@@ -1,3 +1,3 @@
-export { computePowerScore, computeScoringOpportunityBoost, normalizePowerScoreResult } from './scorer';
+export { computePowerScore, computeScoringOpportunityBoost, computeWinProbVarianceScore, normalizePowerScoreResult } from './scorer';
 export * from './types';
 export * from './constants';